chore: Add summary questions endpoint handler.
